% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acm0009_ontology.R
\name{read_acm0009_baseline_fuel_shapes}
\alias{read_acm0009_baseline_fuel_shapes}
\title{Load ACM0009 baseline fuel applicability shape}
\usage{
read_acm0009_baseline_fuel_shapes()
}
\value{
A \code{sh_shape_graph} as returned by \code{\link[shaclR:read_shacl]{shaclR::read_shacl()}}.
}
\description{
Load ACM0009 baseline fuel applicability shape
}
